Remove the Hy version requirement in `setup.py`
We'll continue testing with the latest Hy on GitHub Actions, and I'll continue the practice of pegging a Hy version in each Hyrule release. The point of this change is to allow users of bleeding-edge Hyrule to pick the Hy they want. We leave the problem of choosing a compatible Hy version up to them.